chore: rename mockKv to mockEdgeProvider
